+++ va108xx/examples/embassy/src/bin/async-uart-tx.rs
@@ -70,10 +70,13 @@ async fn main(_spawner: Spawner) {
         led0.toggle();
         led1.toggle();
         led2.toggle();
-        async_tx
-            .write_all(STR_LIST[idx].as_bytes())
-            .await
-            .expect("writing failed");
+        // Safety: We are sending static lifetime slices, and not cancelling the futures.
+        unsafe {
+            async_tx
+                .write_all(STR_LIST[idx].as_bytes())
+                .await
+                .expect("writing failed");
+        }
         idx += 1;
         if idx == STR_LIST.len() {
             idx = 0;