+++ va416xx/va416xx-embassy/docs.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-export RUSTDOCFLAGS="--cfg docsrs --generate-link-to-definition -Z unstable-options"
-cargo +nightly doc --open